languagePrototypes/cel_proto/requirements.txt
cel-python
